+++ src/domain/graphics.rs
@@ -12,14 +12,14 @@ pub enum GraphicDensity {
     /// 180dpi x 180dpi
     Low,
     /// 360dpi x 360dpi
-    Hight,
+    High,
 }
 
 impl From<GraphicDensity> for u8 {
     fn from(value: GraphicDensity) -> Self {
         match value {
             GraphicDensity::Low => 50,
-            GraphicDensity::Hight => 51,
+            GraphicDensity::High => 51,
         }
     }
 }
@@ -28,7 +28,7 @@ impl fmt::Display for GraphicDensity {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             GraphicDensity::Low => write!(f, "180dpi"),
-            GraphicDensity::Hight => write!(f, "360dpi"),
+            GraphicDensity::High => write!(f, "360dpi"),
         }
     }
 }